Introduce pure CSS selectors
A structured selectors without qualifiers.
