+++ src/selector/index.ts
@@ -1,6 +1,7 @@
 /**
  * @module style-producer
  */
+export * from './pure-selector';
 export * from './query';
 export * from './rule-key';
 export * from './selector';
